start function for redis_client and connect_link functions for redis facade
Based on feedback from Paul Rubin, it was clear that the automatic link
created by a call to redis:connect was surprising. The new behavior is to
not link on a call to connect. connect_link functions were added as a pass
through to the respective start_link functions in redis_client.